add login and check is valid user

+++ src/views/Home/FileActions/DeleteAction/DeleteAction.vue
@@ -1,14 +1,17 @@
 <script setup>
-import * as file from '@/models/services/localFile'
-import {defineEmits, defineProps} from 'vue'
-const emit = defineEmits(['delete'])
+import checkIsValidUser from "@/composables/checkIsValidUser";
+import * as file from "@/models/services/localFile";
+import { defineEmits, defineProps } from "vue";
+const emit = defineEmits(["delete"]);
 const props = defineProps({
-  itemId: Number
-})
+  itemId: Number,
+});
 
-function handleDelete() {
-  file.deleteById(props.itemId)
-  emit('delete')
+async function handleDelete() {
+  const isValid = await checkIsValidUser();
+  if (!isValid) return;
+  file.deleteById(props.itemId);
+  emit("delete");
 }
 </script>
 <template>
